+++ gherkin/javascript/src/stream/GherkinStreams.ts
@@ -5,6 +5,7 @@ import ParserMessageStream from './ParserMessageStream'
 import SourceMessageStream from './SourceMessageStream'
 import IGherkinOptions from '../IGherkinOptions'
 import makeGherkinOptions from '../makeGherkinOptions'
+import * as fs from 'fs'
 
 function fromStream(stream: Readable, options: IGherkinOptions) {
   return pipeline(
@@ -36,8 +37,7 @@ function fromPaths(
       const end = pathsCopy.length === 0
       // Can't use pipeline here because of the { end } argument,
       // so we have to manually propagate errors.
-      options
-        .createReadStream(path)
+      fs.createReadStream(path, { encoding: 'utf-8' })
         .on('error', (err) => combinedMessageStream.emit('error', err))
         .pipe(new SourceMessageStream(path))
         .on('error', (err) => combinedMessageStream.emit('error', err))